langchain_MCP_A2A/a2a_agent/pyproject.toml
[project]
name = "a2a-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.9",
    "click>=8.2.1",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.1.5",
    "langchain-mcp-adapters>=0.1.7",
    "langchain[openai]>=0.3.26",
    "langgraph>=0.5.0",
    "mcp[cli]>=1.10.0",
    "pydantic>=2.11.7",
    "uvicorn>=0.34.3",
]
